Moved the petstore docket configuration to its own module
Demonstrates that the issue 1890 is not reproducable.

fixes #1890
